ide: Vastly generalize the 'lookup' query
'lookup' now accepts a context argument, which tells it what kind of information
is thought.  Just like completion, it supports modules, options, and symbols.
Doing this as an extension of 'lookup' rather than a set of new query makes it
easy to extend clients to support the new info.

The change is backwards-compatible (omitting the context causes F* to always
return information about symbols).